!refactor: Extensive refactor
!Perform extensive code refactoring!
!NEW SOLUTION IS NOT COMPATIBLE WITH PREVIOUS SOLUTIONS!

According to the recommendations of the service "code climate"
Function that is the same for all game classes "runGame()" has been reworked,
And moved to the Engine class. Now it is responsible for running
Any game from the game catalog

To achieve the above goal, a public static function "generateTask()"
Has been added to each game class. It changing the game parameters in the Engine class
And creating a task for the player. This function is called from "runGame()"
that is why "runGame()" 'knows' what game parameters it works with
And what game it plays and what it should display on the screen.

To add new games you now only need to:
1 - create a class with a game, implementing
    static public method "generateTask()"
    and method "run()"" in game class
    similar to existing classes
2 - by analogy, add the corresponding lines
    to the functions "printGeneratedTask()" and "runGameByChoice()"